Cart, Checkout,Promo,Sale, Wishlist Fix
if ever I make this on public well don't judge me XD

updated what's in the title

+++ ecom/src/pages/Profile/Address.tsx
@@ -1,31 +1,52 @@
-import React, { useEffect, useState } from "react";
-import { Address } from "../../interfaces/user";
+import React, { useCallback, useEffect, useState } from "react";
+import { AddressInterface } from "../../interfaces/user";
 import ModifyAddress from "./ModifyAddress";
 import axiosInstance from "../../services/axiosInstance";
+import NewAddress from "./NewAddress";
 
 const AddressContent: React.FC<{
-  userAddresses: Address[];
-  onDeleteAddress: (addressId: string) => void;
-  onModifyAddress: (address: Address) => void;
-}> = ({ userAddresses, onDeleteAddress, onModifyAddress }) => {
-  const [addresses, setAddresses] = useState<Address[]>(userAddresses);
+  checkout: boolean;
+  selected?: (address: AddressInterface) => void;
+  toggleClose?: () => void;
+}> = ({ checkout, selected, toggleClose }) => {
+  const [addresses, setAddresses] = useState<AddressInterface[]>([]);
   const [toggleEdit, setToggleEdit] = useState(false);
-  const [currentAddress, setCurrentAddress] = useState<Address | null>(null);
-  useEffect(() => {
-    const sortedAddresses = [
-      ...userAddresses.filter((address) => address.default),
-      ...userAddresses.filter((address) => !address.default),
-    ];
-    setAddresses(sortedAddresses);
-  }, [userAddresses]);
+  const [toggleNewAddress, setToggleNewAddress] = useState(false);
+  const [currentAddress, setCurrentAddress] = useState<AddressInterface | null>(
+    null,
+  );
 
-  const updateAddress = async (updatedAddresses: Address[]) => {
+  const getUserAddress = async () => {
     try {
-      await axiosInstance.post("/update-address", { updatedAddresses });
+      const response = await axiosInstance.get("/get-address");
+      return response.data;
     } catch (error) {
       console.log(error);
     }
   };
+
+  const fetchAddress = useCallback(async () => {
+    const userAddresses = await getUserAddress();
+    const sortedAddresses = [
+      ...userAddresses.filter((address: AddressInterface) => address.default),
+      ...userAddresses.filter((address: AddressInterface) => !address.default),
+    ];
+    setAddresses(sortedAddresses);
+  }, []);
+
+  useEffect(() => {
+    if (toggleNewAddress || toggleEdit) {
+      document.body.classList.add("overflow-hidden");
+    } else {
+      document.body.classList.remove("overflow-hidden");
+    }
+  }, [toggleNewAddress, toggleEdit]);
+
+  useEffect(() => {
+    fetchAddress();
+  }, [fetchAddress]);
+
+ 
 
   const handleSetDefault = async (addressId: string) => {
     try {
@@ -48,24 +69,80 @@ const AddressContent: React.FC<{
     }
   };
 
-  const handleRemove = async (addressId: string) => {
-    if (addressId) {
-      const updatedAddresses = addresses.filter(
-        (address) => !(address._id === addressId),
-      );
-      onDeleteAddress(addressId);
-      setAddresses(updatedAddresses);
-      updateAddress(updatedAddresses);
+  const addNewAddress = async (newAddress: AddressInterface) => {
+    try {
+      const response = await axiosInstance.post("/save-new-address", {
+        newAddress,
+      });
+      if (response.status === 200) {
+        fetchAddress();
+      }
+    } catch (error) {
+      console.log(error);
     }
   };
 
-  const handleEdit = (address: Address) => {
+  const modifyAddress = async (updatedAddress: AddressInterface) => {
+    if (updatedAddress) {
+     try {
+      const updateAddress = await axiosInstance.post("/update-address", { updatedAddress});
+      if (updateAddress.status === 200) {
+        fetchAddress();
+        setToggleEdit(false);
+      }
+    } catch (error) {
+      console.log(error);
+    }
+    }
+  };
+
+  const handleRemove = async (addressId: string) => {
+    try {
+      if (addressId) {
+        const deleteAddress = await axiosInstance.post("/delete-address", {
+          addressId,
+        });
+        if (deleteAddress.status === 200) {
+          fetchAddress();
+        }
+      }
+    } catch (error) {
+      console.error("Failed to remove address", error);
+    }
+  };
+
+  const handleEdit = (address: AddressInterface) => {
     setCurrentAddress(address);
     setToggleEdit(!toggleEdit);
   };
 
+  const handleNewAddress = () => {
+    setToggleNewAddress(!toggleNewAddress);
+  };
+
   return (
     <div className="flex w-full flex-col gap-2">
+      <div className="relative mb-2 flex items-center justify-between border-b-[1px]">
+        <p className="roboto-medium text-lg">Shipping Address</p>
+        <button
+          className="rounded bg-zinc-500 px-2 py-0.5 text-xs text-white hover:bg-zinc-700"
+          onClick={() => handleNewAddress()}
+        >
+          Add New Address
+        </button>
+      </div>
+      {toggleNewAddress && (
+        <>
+          <div className="fixed inset-0 z-50 h-screen w-screen select-none bg-zinc-700/50">
+            <div className="flex h-full w-full flex-col items-center justify-center">
+              <NewAddress
+                toggleClose={handleNewAddress}
+                onAddAddress={addNewAddress}
+              ></NewAddress>
+            </div>
+          </div>
+        </>
+      )}
       {addresses.map((address, index) => (
         <div key={index}>
           {toggleEdit && currentAddress?._id === address._id && (
@@ -74,12 +151,21 @@ const AddressContent: React.FC<{
                 <ModifyAddress
                   toggleClose={() => setToggleEdit(false)}
                   address={currentAddress}
-                  modifiedAddress={onModifyAddress}
+                  modifiedAddress={modifyAddress}
                 ></ModifyAddress>
               </div>
             </div>
           )}
-          <div className="flex flex-col gap-1 rounded border-[1px] p-2 leading-3 sm:max-w-sm">
+          <div
+            className="z-10 flex flex-col gap-1 rounded border-[1px] bg-white p-2 leading-3 hover:cursor-pointer hover:bg-zinc-100 sm:max-w-sm"
+            onClick={(e) => {
+              e.preventDefault();
+              if (checkout) {
+                selected?.(address);
+                toggleClose?.();
+              }
+            }}
+          >
             {address.default && (
               <p className="roboto-medium border-b-[1px] text-xs text-zinc-500">
                 Default Address
@@ -100,24 +186,40 @@ const AddressContent: React.FC<{
             </p>
             <div className="my-1 flex items-center gap-1 text-xs leading-3">
               <button
-                className="rounded-full border-[1px] px-3 py-1 hover:cursor-pointer hover:border-zinc-400 hover:bg-zinc-300"
-                onClick={() => handleEdit(address)}
+                className="z-40 rounded-full border-[1px] px-3 py-1 hover:cursor-pointer hover:border-zinc-400 hover:bg-zinc-300"
+                onClick={(e) => {
+                  e.stopPropagation();
+                  handleEdit(address);
+                }}
               >
                 Modify
               </button>
               <button
-                className="rounded-full border-[1px] px-3 py-1 hover:cursor-pointer hover:border-zinc-400 hover:bg-zinc-300"
-                onClick={() => {
-                  address._id && handleRemove(address._id);
+                className="z-40 rounded-full border-[1px] px-3 py-1 hover:cursor-pointer hover:border-zinc-400 hover:bg-zinc-300"
+                onClick={(e) => {
+                  e.stopPropagation();
+                  if (
+                    window.confirm(
+                      "Are you sure you want to remove this address?",
+                    )
+                  ) {
+                    address._id && handleRemove(address._id);
+                  }
                 }}
               >
                 Remove
               </button>
               {!address.default && (
                 <button
-                  className="ml-auto self-end rounded-full border-[1px] px-2 py-1 leading-3 hover:cursor-pointer hover:border-zinc-400 hover:bg-zinc-300"
-                  onClick={() => {
+                  className="z-40 ml-auto self-end rounded-full border-[1px] px-2 py-1 leading-3 hover:cursor-pointer hover:border-zinc-400 hover:bg-zinc-300"
+                  onClick={(e) => {
+                    e.stopPropagation();
                     address._id && handleSetDefault(address._id);
+
+                    if (checkout) {
+                      const updatedAddress = { ...address, default: true };
+                      selected?.(updatedAddress);
+                    }
                   }}
                 >
                   Set as Default Address
